pyproject.toml
[project]
name = "carousell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "selenium==4.41.0",
    "requests==2.32.5",
    "webdriver-manager==4.0.2",
    "setuptools==82.0.1",
    "seleniumbase==4.47.4",
    "python-dotenv==1.2.2"
]
